+++ crates/node/core/src/cli/config.rs
@@ -38,7 +38,7 @@ pub trait PayloadBuilderConfig {
     /// Maximum number of blobs to include per block (EIP-7872).
     ///
     /// If `None`, defaults to the protocol maximum.
-    fn max_blobs_per_block(&self) -> Option<u8>;
+    fn max_blobs_per_block(&self) -> Option<u64>;
 
     /// Returns the configured gas limit if set, or a chain-specific default.
     fn gas_limit_for(&self, chain: Chain) -> u64 {
